Problem: loading of climate_Muenchenberg92_98.csv in cmf_setup.py doesn't  work:
Traceback (most recent call last):
  File "C:\Users\Pooder\Documents\PMFCMF30062012\sensitivity_analysis\sensitivity_analysis_I.py", line 178, in <module>
    c.load_meteo(rain_factor=1.)
  File "C:\Users\Pooder\Documents\PMFCMF30062012\sensitivity_analysis\cmf_setup.py", line 176, in load_meteo
    [float(col) for col in columns[0:]]):    
ValueError: could not convert string to float:

+++ sensitivity_analysis/Atmosphere.py
@@ -90,7 +90,7 @@ class ClimateStation: #!!!
         csvfile.readline()
         for line in csvfile:
             columns = line.split(';')
-            Wetter[columns[0]] = float(columns[value])
+            Wetter[columns[0]] = double(columns[value])
     def __call__(self,t, time_step):
         t += time_step
         return t
